# app/core/swagger/swagger.yaml
basePath: /api
definitions:
  audit.AuditAction:
    enum:
    - CREATE
    - UPDATE
    - DELETE
    - LIST
    type: string
    x-enum-varnames:
    - ActionCreate
    - ActionUpdate
    - ActionDelete
    - ActionList
  audit.AuditModel:
    properties:
      action:
        allOf:
        - $ref: '#/definitions/audit.AuditAction'
        description: Action is the operation that was performed (CREATE, UPDATE, DELETE,
          LIST).
      actor:
        description: |-
          Actor is the identity that triggered the operation (e.g. a username or
          certificate CN). Defaults to "system" when no actor is available.
        type: string
      details:
        description: |-
          Details holds any additional context about the operation
          (e.g. the payload snapshot or a short description).
        type: string
      entity:
        description: Entity identifies the domain object being audited (e.g. "Metadata").
        type: string
      entity_id:
        description: |-
          EntityID is the primary-key value of the affected record.
          Empty string means the operation targeted the entire collection (e.g. LIST).
        type: string
      id:
        description: ID is an auto-incremented sequence number assigned at insertion
          time.
        type: integer
      occurred_at:
        description: OccurredAt is the UTC timestamp of when the operation was recorded.
        type: string
    type: object
  metadata.MetadataModel:
    properties:
      asset_id:
        type: integer
      created_at:
        description: Dates are stored as RFC3339 strings, matching the chaincode's
          on-chain representation.
        type: string
      created_by:
        type: string
      deleted_at:
        type: string
      deleted_by:
        type: string
      id:
        type: integer
      name:
        type: string
      owner:
        type: string
      patient_id:
        type: integer
      rights:
        type: string
      terms_of_access:
        type: string
      updated_at:
        type: string
      updated_by:
        type: string
      value:
        type: string
      version:
        type: string
      zkp_proof:
        type: string
    type: object
  utils.ErrorResponse:
    properties:
      message:
        example: An error occurred
        type: string
      success:
        example: false
        type: boolean
    type: object
  utils.SuccessResponse:
    properties:
      data: {}
      message:
        example: Operation completed successfully
        type: string
      success:
        example: true
        type: boolean
    type: object
host: localhost:8080
info:
  contact:
    name: Silas Silva
    url: https://github.com/silvasilas99/entruster
  description: REST API for registering and querying healthcare metadata on a Hyperledger
    Fabric blockchain.
  license:
    name: MIT
  termsOfService: http://swagger.io/terms/
  title: Entruster API
  version: "1.0"
paths:
  /metadata/:
    get:
      description: Evaluates GetAllMetadataFromNetwork on the Fabric ledger and returns
        every stored metadata asset.
      parameters:
      - description: Filter by patient ID
        in: query
        name: patient_id
        type: string
      - description: Filter by asset ID
        in: query
        name: asset_id
        type: string
      - description: Start date (created_at)
        in: query
        name: from
        type: string
      - description: End date (created_at)
        in: query
        name: to
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/utils.SuccessResponse'
            - properties:
                data:
                  items:
                    additionalProperties: true
                    type: object
                  type: array
              type: object
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
      summary: List all metadata assets
      tags:
      - metadata
    post:
      consumes:
      - application/json
      description: Submits a new metadata asset to the Hyperledger Fabric ledger.
        The asset ID is auto-generated by the chaincode.
      parameters:
      - description: Metadata payload (patient_id and asset_id must be numeric strings)
        in: body
        name: metadata
        required: true
        schema:
          $ref: '#/definitions/metadata.MetadataModel'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            allOf:
            - $ref: '#/definitions/utils.SuccessResponse'
            - properties:
                data:
                  additionalProperties:
                    type: string
                  type: object
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
      summary: Register metadata on the blockchain
      tags:
      - metadata
  /metadata/{id}:
    delete:
      description: Submits a DeleteMetadataById transaction to the Fabric ledger,
        removing the asset permanently.
      parameters:
      - description: Asset ID (numeric string)
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/utils.SuccessResponse'
            - properties:
                data:
                  additionalProperties:
                    type: string
                  type: object
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
      summary: Delete metadata by ID
      tags:
      - metadata
    get:
      description: Evaluates GetMetadataById on the Fabric ledger and returns the
        matching metadata asset.
      parameters:
      - description: Asset ID (numeric string, e.g. \
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/utils.SuccessResponse'
            - properties:
                data:
                  $ref: '#/definitions/metadata.MetadataModel'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
      summary: Get metadata by ID
      tags:
      - metadata
    put:
      consumes:
      - application/json
      description: Submits an UpdateMetadataById transaction to the Fabric ledger
        with the provided field values.
      parameters:
      - description: Asset ID (numeric string)
        in: path
        name: id
        required: true
        type: string
      - description: Updated metadata fields
        in: body
        name: metadata
        required: true
        schema:
          $ref: '#/definitions/metadata.MetadataModel'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/utils.SuccessResponse'
            - properties:
                data:
                  additionalProperties:
                    type: string
                  type: object
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
      summary: Update metadata by ID
      tags:
      - metadata
  /metadata/{id}/auditory:
    get:
      description: Returns the full immutable history of the asset from the audit
        service.
      parameters:
      - description: Asset ID (numeric string)
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/utils.SuccessResponse'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/audit.AuditModel'
                  type: array
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
      summary: Get audit trail for a metadata asset
      tags:
      - metadata
schemes:
- http
swagger: "2.0"